% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat-util.R
\name{create_template}
\alias{create_template}
\title{Create a raster template}
\usage{
create_template(lon, lat, crs, res_km)
}
\arguments{
\item{lon}{\verb{[dbl(2)]} Longitudinal extent, xmin and xmax.}

\item{lat}{\verb{[dbl(2)]} Latitudinal extent, ymin and ymax.}

\item{crs}{\verb{[chr(1)]} proj4 string describing projection of raster template.}

\item{res_km}{\verb{[dbl(1)]} Resolution of raster template in km.}
}
\value{
\verb{[terra::SpatRaster]}
}
\description{
Create a common grid across variables for data extraction and prediction.
}
\examples{
# A CRS appropriate for the California Current
# Specifically, an oblique Lambert azimuthal equal area projection centered
# on 36N, 122W
crs <- "+proj=laea +lat_0=36 +lon_0=-122 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"
# A raster template roughly covering the California Current from San Diego to
# Arcata at a 10km horizontal resolution.
r <- create_template(
  lon = c(-127, -117),
  lat = c(32, 40),
  crs = crs,
  res_km = 10
)
r
}
